Sort buckets with merge
After receiving buckets, they can be sorted with a multiway merge rather than a
full sort since they come in ordered chunks.
However, this appears to be slower than just running std::sort, at least on
smaller inputs.
I'm pushing this anyway (to a separate branch) just to have the code. We may find
that it's faster on large inputs. (But first we need to change our
implementation to support large inputs, and we also need to find large test
data...)