Fix sizing issue in ssort
In the "redistribute" function, there's a point where I calculate the positions
of elements in a bucket relative to all the elements across all processors. I
can use size_t here instead of int.
This means that the number of elements of data across all the processors can exceed
2^31 (although the number of elements held on any particular processor still
needs to be within 2^31).