Use smaller datatypes
Change uint64_t to uint32_t. Turns out that this implementation is pretty slow,
and there's no way that we're going to be able to run it quickly on large
datasets. And if we aren't going to test on large datasets, we don't have to
worry about integer overflow.
Using smaller datatypes makes communication a little more efficient.